Migrating de Flood Fill algorithm implementation to C.
I've started recoding the Flood Fill algorithm to C, and creating a python wrapper so
Painthon can use it.

Right now, I have the wrapper configured properly, and a C skeletton which works. The
algorithm itself, though, is not implemented. The skeletton performs a substitution
of all the pixels in the image (fills the whole image with the given replacement color).