Decoupling painthon's state from GUI.
I've decided to decouple a little bit the UI from Painthon itself. Now, Painthon's class provides a set of functions to the current state of the draw, and the UI is just a window to this state.

Save and Save As functions (and probably Open, New, etc) do not work properly yet, but I hope they will soon.

This refactor improves code and will hopelly simplify the creation of a new, different UI.